+++ include/granit/tools/shader_tools.h
@@ -5,6 +5,7 @@
 #define GRANIT_SHADER_TOOLS_H_
 
 #include <granit/tools/shader_compiler.h>
+#include <granit/tools/shader_library_builder.h>
 #include <granit/tools/shader_object_builder.h>
 #include <granit/tools/shader_reflection.h>